requirements.txt
tinydb
flask
flask-sqlalchemy
flask-migrate
gitpython
bcrypt